Clean up
Source: https://github.com/msupply-foundation/open-msupply/commit/a864d370b674a1e5f2d7261b2e88e4292436b1bc

+++ server/graphql/general/src/queries/tests/item_stats.rs
@@ -50,15 +50,15 @@ mod tests {
                     "stats": {
                         "averageMonthlyConsumption":  test_item_stats::item1_amc_3_months(),
                         "availableStockOnHand":  test_item_stats::item_1_soh(),
-                        "availableMonthsOfStockOnHand": test_item_stats::item_1_soh() as f64 / test_item_stats::item1_amc_3_months()
+                        "availableMonthsOfStockOnHand": test_item_stats::item_1_soh() / test_item_stats::item1_amc_3_months()
                     }
                 },
                 {
                     "id": &test_item_stats::item2().id,
                     "stats": {
                         "averageMonthlyConsumption":  test_item_stats::item2_amc_3_months(),
-                        "availableStockOnHand":  test_item_stats::item_2_soh() ,
-                        "availableMonthsOfStockOnHand": test_item_stats::item_2_soh() as f64 / test_item_stats::item2_amc_3_months()
+                        "availableStockOnHand":  test_item_stats::item_2_soh(),
+                        "availableMonthsOfStockOnHand": test_item_stats::item_2_soh() / test_item_stats::item2_amc_3_months()
                     },
                 }]
             }
